split splash into file
